Add the modules `quadtools` and `quadutils`.
The implemented "Quad" classes help to define submodels that perform numerical integrations via Gauß-Lobatto quadrature. There is much room left for optimisation, but it at least works. We show this for the toy method `Calc_DVEq_V1` of base model `wland`, which we can use for later optimisations.

One simple improvement would be to remember the number of nodes required in the last integration run, to not always start with the lowest number of nodes allowed.